+++ externs/protocol_externs.js
@@ -6564,6 +6564,30 @@ Protocol.TracingDispatcher.prototype.dataCollected = function(value) {};
  * @param {Protocol.Tracing.StreamCompression=} opt_streamCompression
  */
 Protocol.TracingDispatcher.prototype.tracingComplete = function(opt_stream, opt_streamCompression) {};
+Protocol.Testing = {};
+
+
+/**
+ * @constructor
+*/
+Protocol.TestingAgent = function(){};
+
+/**
+ * @param {string} message
+ * @param {string=} opt_group
+ * @return {!Promise<undefined>}
+ */
+Protocol.TestingAgent.prototype.generateTestReport = function(message, opt_group) {};
+/** @typedef {!{message: string, group: (string|undefined)}} */
+Protocol.TestingAgent.GenerateTestReportRequest;
+/** @typedef {Object|undefined} */
+Protocol.TestingAgent.GenerateTestReportResponse;
+/**
+ * @param {!Protocol.TestingAgent.GenerateTestReportRequest} obj
+ * @return {!Promise<!Protocol.TestingAgent.GenerateTestReportResponse>} */
+Protocol.TestingAgent.prototype.invoke_generateTestReport = function(obj) {};
+/** @interface */
+Protocol.TestingDispatcher = function() {};
 Protocol.Console = {};
 
 
@@ -8339,6 +8363,12 @@ Protocol.TargetBase.prototype.tracingAgent = function(){};
  * @param {!Protocol.TracingDispatcher} dispatcher
  */
 Protocol.TargetBase.prototype.registerTracingDispatcher = function(dispatcher) {}
+/** @return {!Protocol.TestingAgent}*/
+Protocol.TargetBase.prototype.testingAgent = function(){};
+/**
+ * @param {!Protocol.TestingDispatcher} dispatcher
+ */
+Protocol.TargetBase.prototype.registerTestingDispatcher = function(dispatcher) {}
 /** @return {!Protocol.ConsoleAgent}*/
 Protocol.TargetBase.prototype.consoleAgent = function(){};
 /**