+++ changelog.md
@@ -1,5 +1,11 @@
 
 
+## Roll protocol to r586417
+###### _2018-08-27 14:15:56_ | Diff: [fbe2ce1...cafc591](https://github.com/ChromeDevTools/devtools-protocol/compare/fbe2ce1...cafc591)
+#### `Page`: new command
+* [`Page.generateTestReport`](https://chromedevtools.github.io/devtools-protocol/tot/Page/#method-generateTestReport)
+
+
 ## Roll protocol to r585632
 ###### _2018-08-23 15:15:44_ | Diff: [d4361d7...fbe2ce1](https://github.com/ChromeDevTools/devtools-protocol/compare/d4361d7...fbe2ce1)
 #### `Browser`: new commands